requirements.txt
customtkinter
esptool
pyserial
packaging
pillow
